requirements.txt
asgiref==3.8.1
Django==5.1
psycopg2==2.9.10
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
